feat: move data fetches from App.tsx to custom hook useFetchData.ts
